Moved files service API endpoint functions to separate files
